Revert "Well units (#167)"
This reverts commit 6adfe15fdcbc943704de01af6227283a7983daeb.
