+++ ResSimpy/Node.py
@@ -5,8 +5,7 @@ from typing import Optional
 
 from ResSimpy.Enums.UnitsEnum import UnitSystem
 from ResSimpy.DataObjectMixin import DataObjectMixin
-from ResSimpy.Units.AttributeMappings.AttributeMappingBase import AttributeMapBase
-from ResSimpy.Units.AttributeMappings.NetworkUnitAttributeMapping import NetworkUnits
+from ResSimpy.Units.AttributeMapping import NetworkNodesConnections, AttributeMapBase
 
 
 @dataclass
@@ -20,4 +19,4 @@ class Node(DataObjectMixin, ABC):
     @property
     def attribute_to_unit_map(self) -> AttributeMapBase:
         """Returns the attribute to unit map for the constraint."""
-        return NetworkUnits()
+        return NetworkNodesConnections()
