+++ ResSimpy/NodeConnection.py
@@ -1,11 +1,9 @@
-"""Base class object for storing data related to nodeconnections."""
 from __future__ import annotations
 from abc import ABC
 from dataclasses import dataclass
 from typing import Optional
 from ResSimpy.DataObjectMixin import DataObjectMixin
-from ResSimpy.Units.AttributeMappings.AttributeMappingBase import AttributeMapBase
-from ResSimpy.Units.AttributeMappings.NetworkUnitAttributeMapping import NetworkUnits
+from ResSimpy.Units.AttributeMapping import AttributeMapBase, NetworkNodesConnections
 
 
 @dataclass
@@ -21,4 +19,4 @@ class NodeConnection(DataObjectMixin, ABC):
     @property
     def attribute_to_unit_map(self) -> AttributeMapBase:
         """Returns the attribute to unit map for the constraint."""
-        return NetworkUnits()
+        return NetworkNodesConnections()
