+++ ResSimpy/Target.py
@@ -4,8 +4,7 @@ from abc import ABC
 from dataclasses import dataclass, field
 from typing import Optional
 from ResSimpy.DataObjectMixin import DataObjectMixin
-from ResSimpy.Units.AttributeMappings.AttributeMappingBase import AttributeMapBase
-from ResSimpy.Units.AttributeMappings.NetworkUnitAttributeMapping import NetworkUnits
+from ResSimpy.Units.AttributeMapping import AttributeMapBase, NetworkNodesConnections
 
 
 @dataclass
@@ -34,4 +33,4 @@ class Target(DataObjectMixin, ABC):
     @property
     def attribute_to_unit_map(self) -> AttributeMapBase:
         """Returns the attribute to unit map for the WellConnection."""
-        return NetworkUnits()
+        return NetworkNodesConnections()
