+++ ResSimpy/Units/Units.py
@@ -257,16 +257,6 @@ class Pressure(UnitDimension):
     metric_atm = 'atm'
 
 
-class DeltaPressure(UnitDimension):
-    """Units for pressure."""
-    english = 'psi'
-    metric = 'kPa'
-    metkgcm2 = 'kg/cm2'
-    metbar = 'bars'
-    lab = 'psi'
-    metric_atm = 'atm'
-
-
 class ReservoirRates(UnitDimension):
     """Units for reservoir rates."""
     english = 'RB/day'
@@ -448,30 +438,9 @@ class HeatTransfer(UnitDimension):
 
 
 class Dimensionless(UnitDimension):
-    """For attributes which do not have a dimension."""
     english = ''
     metric = ''
     metkgcm2 = ''
     metbar = ''
     lab = ''
     metric_atm = ''
-
-
-class PermeabilityThickness(UnitDimension):
-    """Permeability thickness units."""
-    english = 'md*ft'
-    metric = 'md*m'
-    metkgcm2 = 'md*m'
-    metbar = 'md*m'
-    lab = 'md*m'
-    metric_atm = 'md*m'
-
-
-class NonDarcySkin(UnitDimension):
-    """Non darcy skin units."""
-    english = 'day/MSCF'
-    metric = 'day/SM3'
-    metkgcm2 = 'day/SM3'
-    metbar = 'day/SM3'
-    lab = 'hour/stcc'
-    metric_atm = 'day/SM3'
