+++ ResSimpy/WellConnection.py
@@ -4,8 +4,7 @@ from typing import Optional
 
 from ResSimpy.Enums.UnitsEnum import UnitSystem
 from ResSimpy.DataObjectMixin import DataObjectMixin
-from ResSimpy.Units.AttributeMappings.AttributeMappingBase import AttributeMapBase
-from ResSimpy.Units.AttributeMappings.NetworkUnitAttributeMapping import NetworkUnits
+from ResSimpy.Units.AttributeMapping import AttributeMapBase, NetworkNodesConnections
 
 
 class WellConnection(DataObjectMixin, ABC):
@@ -27,4 +26,4 @@ class WellConnection(DataObjectMixin, ABC):
     @property
     def attribute_to_unit_map(self) -> AttributeMapBase:
         """Returns the attribute to unit map for the WellConnection."""
-        return NetworkUnits()
+        return NetworkNodesConnections()
