+++ ResSimpy/Wellbore.py
@@ -6,8 +6,7 @@ from typing import Optional
 
 from ResSimpy.Enums.UnitsEnum import UnitSystem
 from ResSimpy.DataObjectMixin import DataObjectMixin
-from ResSimpy.Units.AttributeMappings.AttributeMappingBase import AttributeMapBase
-from ResSimpy.Units.AttributeMappings.NetworkUnitAttributeMapping import NetworkUnits
+from ResSimpy.Units.AttributeMapping import AttributeMapBase, NetworkNodesConnections
 
 
 @dataclass
@@ -22,4 +21,4 @@ class Wellbore(DataObjectMixin, ABC):
     @property
     def attribute_to_unit_map(self) -> AttributeMapBase:
         """Returns the attribute to unit map for the WellConnection."""
-        return NetworkUnits()
+        return NetworkNodesConnections()
