+++ tests/Units/test_unit_mappings.py
@@ -4,14 +4,9 @@ from ResSimpy.Enums.UnitsEnum import UnitSystem
 from ResSimpy.Nexus.DataModels.Network.NexusConstraint import NexusConstraint
 from ResSimpy.Nexus.DataModels.Network.NexusNode import NexusNode
 from ResSimpy.Nexus.DataModels.Network.NexusNodeConnection import NexusNodeConnection
-from ResSimpy.Nexus.DataModels.Network.NexusTarget import NexusTarget
-from ResSimpy.Nexus.DataModels.Network.NexusWellConnection import NexusWellConnection
-from ResSimpy.Nexus.DataModels.Network.NexusWellbore import NexusWellbore
-from ResSimpy.Nexus.DataModels.Network.NexusWellhead import NexusWellhead
-from ResSimpy.Nexus.DataModels.NexusCompletion import NexusCompletion
 from ResSimpy.Units.Units import Area
-from ResSimpy.Units.AttributeMappings.ConstraintUnitAttributeMapping import ConstraintUnits
-from ResSimpy.ISODateTime import ISODateTime
+from ResSimpy.Units.AttributeMapping import ConstraintUnits
+
 
 @pytest.mark.parametrize('unit_system, expected_result', [
     (UnitSystem.ENGLISH, 'ft2'),
@@ -30,7 +25,6 @@ def test_unit_system_enum_to_variable(unit_system, expected_result):
     result = unit_dimension.unit_system_enum_to_variable(unit_system=unit_system)
     # Assert
     assert result == expected_result
-
 
 @pytest.mark.parametrize('attribute, unit_system, expected_result', [
     ('max_surface_water_rate', UnitSystem.ENGLISH, 'STB/day'),
@@ -57,47 +51,26 @@ def test_get_unit(attribute, unit_system, expected_result):
     assert result == expected_result
 
 
-def test_get_unit_upper():
-    """Tests the get_unit method."""
-    # Arrange
-    unit_dimension = ConstraintUnits()
-
-    # Act
-    result = unit_dimension.get_unit_from_attribute(unit_system=UnitSystem.ENGLISH,
-                                                    attribute_name='max_surface_water_rate', uppercase=True)
-    # Assert
-    assert result == 'STB/DAY'
-
-
 def test_get_unit_error():
     """Tests the get_unit method."""
     # Arrange
     unit_dimension = ConstraintUnits()
 
     # Act
-    with pytest.raises(AttributeError) as ae:
+    with pytest.raises(AttributeError):
         unit_dimension.get_unit_from_attribute(unit_system=UnitSystem.ENGLISH, attribute_name='not_an_attribute')
-    assert str(ae.value) == 'Attribute not_an_attribute not recognised and does not have a unit definition'
 
 
-@pytest.mark.parametrize('data_object, attribute, expected_result, upper', [
-    (NexusConstraint, 'max_surface_water_rate', 'STB/day', False),
-    (NexusNode, 'depth', 'ft', False),
-    (NexusNodeConnection, 'diameter', 'in', False),
-    (NexusWellbore, 'measured_depth_in', 'ft', False),
-    (NexusCompletion, 'angle_a', 'degrees', False),
-    (NexusCompletion, 'angle_a', 'DEGREES', True),
-    (NexusTarget, 'calculation_method', '', True),
-    (NexusWellhead, 'dp_add', 'psi', False),
-    (NexusWellConnection, 'dt_add', 'DEGREES F', True),
+@pytest.mark.parametrize('data_object, attribute, expected_result', [
+    (NexusConstraint, 'max_surface_water_rate', 'STB/day'),
+    (NexusNode, 'depth', 'ft'),
+    (NexusNodeConnection, 'diameter', 'in'),
 ])
-def test_get_unit_for_attribute(mocker, data_object, attribute, expected_result, upper):
+def test_get_unit_for_attribute(data_object, attribute, expected_result):
     """Write a test to check that the DataObjectMixin.get_unit_for_attribute method works as expected."""
     # Arrange
-    'patch out convert_to_iso from the ISODateTime module as it is not needed for this test'
-    mocker.patch.object(ISODateTime, 'convert_to_iso', return_value=ISODateTime(2021, 1, 1))
-    dataobj = data_object({})
+    constraint = data_object({})
     # Act
-    result = dataobj.get_unit_for_attribute(attribute_name=attribute, unit_system=UnitSystem.ENGLISH, uppercase=upper)
+    result = constraint.get_unit_for_attribute(attribute_name=attribute, unit_system=UnitSystem.ENGLISH)
     # Assert
     assert result == expected_result
